Uniting all WFC backend code under Fannie
legacy is mostly old interfaces that people are used to.
Most of these already have newer equivalents in fannnie,
but the old versions all heed the centralized config
at least. Or mostly.

cache is meant to deal with an HTML caching db (for
reports, mostly). Not really configurable right now.

reports/Store-Specific is for what it sounds like:
specialized reporting for a location.