Replaced pinghost(hostname) with pingport(hostname, dbms)
It connects to the host on the default port for the database
(or the one specified by hostname, e.g. 127.0.0.1:3306) then
immediately disconnects again. In theory this should be an
improvement because it verifies that the database is [probably]
available as well as whether the host is online.

Requires PHP socket extension, but that's typically available
by default