I started revising the lane install to be more like Fannie with separate files defining each table and view. I also removed a few unused or rarely used structures.
promomsgs and promoMsgsView are both gone. They were only used in
the "new" receipt views (which may go away as well eventually) and
the customReceipt table provides the same functionality.

I moved chargecode and chargeCodeView into a plugin called
WedgeStaffCharge. As far as I know no one is using this feature
and it doesn't need to be present in the default install. I put the
corresponding Parser in the plugin as well as one of the
deprecated PrehLib functions.