pos background image as symlink, default to empty
The PoS background image was defaulting or being changed to the one
for whatever coop most recently committed, at the moment, People's.
Changed is4c.gif to a symlink, default to an empty image.
symlink should perhaps be gitignored.
I forgot to put the README in this commit.  It will contain:
is4c.gif is used as a background image on the main PoS page.
peoples_coop_logo.gif is an example of a suitable image.
The distributed is4c.gif is symbolic link to, by default,
 a tiny, empty image; a placeholder.
 If you have a better image, change the link to that:
 $ rm is4c.gif
 $ ln -s my_better_image.gif is4c.gif