Class-based Report Refactors: * RecentSales * SalePerformance * StockSummary * SuspensionHistory * TenderInOut * Trends
I extended the ProductMovement report to search non-numeric
UPCs differently. I use this for tracking POS logging records.

SQLManager has a new week() method for getting the week #.

SalesToday is NOT an instance of FannieReportPage although it is
class based. There's not that much data and it's only live so
export makes no sense. I added a line graph using d3.js.