Safely Renaming a Column in custdata, pt 3 ====================================== Lane Code Cleanup ====================================== Easy. The lane only access the limit value via the memchargebalance view which has already been updated. So this just removes stale references.
PrehLib and the gui-modules all issue the same
query to get a record from custdata. The query
requests the MemDiscountLimit column but never
does anything with it. No reason to have the
column name there.

CoreState looks up the limit value in its
crash recovery function but does not do anything
useful with it. Again no reason to have the column
name there.