FannieReportPage header counter
$header_index for pointing to two-dimensional $report_headers
instead of using $multi_count
I've initialized this for each report in multi-report mode but
don't have a report that uses both multiple header sets and multiple
reports to test it with.

Put out $report_description as-is if it is in HTML tags,
otherwise prefix <br /> as before.

assign_headers() as a placeholder for $report_headers assignment in
multi-report-mode as alternative to doing it in calculate_footers()
I've re-coded GeneralDayReport.php to test this, but don't plan
to commit that version.

Increment $multi-counter for xls multi-reports

Use strip_tags() on csv and xls report bodies.

Right-alignment of numbers a bit more robust:
 o embedded commas OK
 o spaces after leading $
 o spaces before trailing %
uses preg_match, with exhaustive notes.
On a page with 1,000 rows with 7 uses of preg_match in each row
 preg_match was 7/1000 sec faster than the less-robust string
operations.

Remove $,%,',' from xls numbers

Improve xls header handling

Moved un-formatting for csv into excelFormat().
Also used on xls, but not certain this is wanted.

In drawPage() restricted output of JS and CSS extras to HTML reports.