Add table SpecialDeptMap to replace configuration setting of the same name. Fitting the setting into the parameters table is not feasible. This approach allows central management of the setting by syncing tables.
The configuration UI will migrate existing settings into the table
and REMOVE the setting from ini.php. The code will use the setting
from $CORE_LOCAL if present so that nothing will break if the code
is updated but the user has not yet visited the configuration UI page
and migrate settings.