Cleaning up spaghetti code related to setting member number on a transaction.
PrehLib::setMember now just takes a CardNo and a
personNum value. The 3rd argument with a row of special database
values is no longer used. The method can query the database itself.

PrehLib::memberID() no longer takes an unnecessary trip through
the member list page. The old behavior when exactly one record
matched would:
1. Redirect to memlist.php
2. Repeat the search
3. Find exactly one record, again
4. Redirect back to pos2.php
This is ridiculous and makes the logic on the search page even
more tortured than it needs to be.