Use PHP_CodeSniffer if available during Fannie Plugin unit test
This will generate *warnings* for the following conditions:
  - class part of namespace
  - method name is not camelCase
Warnings can be ignored at this point; they stem from the CORE
standard being built off the PSR1 coding standard.

The test will generate *errors* if the file has unacceptable
side effects (or if the code analyzer doens't understand a
sequence of tokesn). The following side effects are permitted:
  - include() on config.php
  - include_once() on config.php
  - require() on config.php
  - require_once() on config.php
  - FannieDispatch::conditionalExec();
  - An "if !class_exists" block
  - An "if !function exists" block
  - An "if" construct checking __FILE__ against $_SERVER['PHP_SELF']
Note that any of the allowed "if" blocks *must* include
braces { } or the analyzer will mark them as errors.