Database.php member_subtotal parameter test
CoreLocal::get("member_subtotal") should not test the data type. The actual value is 1/0, not true/false.
Related commit to PrehLib.php

Also, I think the --COMMENTS should be removed.  They aren't the way we do this. I used it when I was new here and wanted to make clear what I was changing.