Make miscellaneous javascript libs match overall bootstrap theme better
Added an alternate stylesheet for Chosen which drives the
type-to-filter <select>s on the item editor. Credit:
https://github.com/alxlit/bootstrap-chosen

Added tablesorter-2.22.1 side by side with the existing one
and did some prep work on FannieReportPage to use the newer
tablesorter. Prep work is still commented out. Issues I see
in switching to the newer tablesorter w/ bootstrap theme:
* Up/down glyphicons do not appear is column headers. Unclear why.
* The sorted column is not highlighted separately from other headers
* Overall text size is larger. Possibly nice for viewing but less
  ideal for printing.
* Paths change between versions. May need to restructure 2.22.1
  so a simple symlink switch between versions won't break non-report
  pages that use tablesorter.
* Reports need print CSS. May help with the text-size issue.