ProductList support multiDept without superdept
Currently the multiDept search returns nothing if a superdept isn't chosen because it get gets into a branch that assumes a range of depts (that branch does not require a superdept).  I don't think there's a need for the superdept if multiDept is preferred (The superdept selection makes the subdepts available, which the multiDept choice doesn't, but perhaps confusingly since it lists all the subdepts in the super, not just those for the chosen depts). The snippet at 594-600 is just a copy of the code from handling multiDept if there is a superdept.

I also suggest here (line 630) a distinction in handling between a failed search and one that returns nothing.  The failure notice would have helped me figure out the bug in subdept search that I reported earlier today.  Having a "Back" link in the message would be a nice touch.