Item Editor subdepts default
Currently if a department has subdepts and `products.subdept = 0` the editor does not offer the 0=None option and assigns the alphabetically first subdept on Update.  The proposed change supports the 0=None option, including `SELECTED` if it is the current value.  The chained selectors in the page still work as they did with this change.

On a related issue, the behaviour I would like when the department is changed and then changed back to the original value is for the original subdept to also be restored.  Currently the current subdept is lost if dept is changed.  I don't know how this might be done.