Show character set info on lane debug page. This simply probes the system to determine the current values.
First it issues an HTTP request to itself and displays the
request headers. Any charset specified in the HTTP headers will
often (read: browser dependent) override a charset that's embedded
in the page via <meta> tags. The HTTP charset dictates how things
are rendered on screen as well as what encoding is used when
submitting forms.

Next it connects to both local and server databases and gets the
relevant charset settings for the connection. These indicate
what charset PHP should use when submitting data to the database
as well as what encoding the database will use when responding
to queries.

Whether the webserver and database connection agree on a charset
is vastly more important than what charset a given table or
column in the database uses. If the webserver and database
connection disagree values may be saved incorrectly and/or
displayed incorrectly.