This would work fine as-is, but to address questions:
Using "[]" at the end of a form field's name attribute
makes that field an array of values. These values get
listed in the order they appear, so in a table you can
have an array of IDs, an array of tax values, array
of WIC values, etc and they'll line up correctly w/o
embedding an identifier in every field.

Using POST instead of GET sometimes works better for
large data sets. Browsers typically have a max URL length
limit so there's only so much data you can tack on after
the ? (often there's a POST size limit too somewhere in the
stack but it's larger).

Updating the product records directly incurs a lot less
overhead but requires extra steps to mimic everything the
model does automatically. The first is manually setting
products.modified. The second is creating the prodUpdate
records to correspond to the change.