Optional asynchronous UDP
The implementation for receiving messages over UDP has
always been synchronous. The driver listens for a message
over UDP, passes it to serial port handler(s), **waits for
the handler(s) to finish dealing with the message**, and
then resumes listening for messages over UDP.

This is problematic with ActiveX controls that receive
an external message and take some action that **blocks**
on user input. If, for example, a signature is requested,
the ActiveX control blocks waiting for the signature and the
driver cannot receive any additional messages, such as
"cancel signature request", until the ActiveX control
returns a value.

Because using Begin*/End* functions didn't work correctly
in some instances on linux I'm making this an optional toggle
(as well as because it's brand new and not production tested).
By default UDP communication remains synchronous. Adding a
"asyncUDP" key with value true to ini.json switches on the
new asynchronous mode.

This commit also changes RBA_Stub because the new potential
for concurrent messages can cause issues with ownership of
the single active connection to a device.