Expose dtransactions.date_id in dlog views
This column may or may not exist in the underling table. If
it does, the dlog view just passes the value through. If it
does not, the dlog view uses datetime functions to calculate
the appropriate value.

In the long run, adding the column and indexing it may be
beneficial if lots of reporting JOINs rely on the column, but
doing it this way leaves changing or not changing bigArchive
to the DBA's discretion.