Generic support for jquery.numpad
If touchscreen setting is enabled, jquery.numpad is
included and inputs with css class "touchtype" will display
a numpad on touch/click. Inputs with additional css class
"submitable" will submit the nearest form when the numapd
is closed; otherwise focus returns to the current input.

jquery.numpad is slightly modified to include a 00 key
and pass the original input element to the numpad close
callback function.