Databar coupon changes
The scale itself attaches a RX prefix to databar values. The
driver also prefixes GS1~ so values passed to POS will have a
full prefix of GS1~RX.

Testing so far is based off coupons with a $ off value and a
single purchase requirement. The spec here is pretty complex.
Each coupon has:
* A standalone value code
* At least one but up to three requirement blocks
* Each requirement block also has a value that may relate to
  the coupons value or to a quantity requirement depending on
  the type of the particular requirement block
* There are several other modifiers that can follow. Expiration
  date is the only one I've verified gets parsed correctly
* Everything is laid out sequentially meaning errors in the
  parsing of optional sections could cause cascading problems
  in later fields. Sourcing actual coupons that contain all the
  various optional sections for testing seems very difficult.