Fix erasing too-deep sign price
I find that the the too-deep (line breaks) price doesn't quite get erased, so I moved the vertical coordinate up a bit. A hack admittedly. I don't know why the cursor position when `MultiCell()` was called for the price and the vertical parameter for the overwrite `Rect()` are not the same.
I put a picture in #866.