InstaCart revisions
Moved the file-generator to a separate class. Added
a 2nd of these for the newer version 3.0 file format.
The submission task now just calls a file-generator. I also
added a report for viewing samples of what the submission
will look like. The HTML version of the report is cut off
to prevent browser hangs/crashes with lots of data but
downloading the report will give the full file.