Allow default value true in BasicModel columns
Boolean true here has a special meaning. The column
must have a non-NULL default value but the model
will not try to manage this attribute after the column
has been created. In effect this means the table will
be created with a non-NULL default but the user can
manually change that default value to something else
and the update system won't try to change it back.