Change RBA_Stub stop mechanism
Previously stopping would just close the serial port
expecting the Read() thread would throw an exception,
catch it, and exit. But Read() actually starts a separate
background thread for asynchronous operation. If this
background thread throws the exception it can't be
caught without tearing down the whole application. So
now the stop action just changes the boolean SPH_Running
flag and waits for the Read() thread to exit cleanly.
This inherently adds a <= 500ms delay waiting for the
current read operation on the serial port to time out.

The Read() thread *also* now waits if its byte buffer
isn't empty. In that case it will continue reading bytes
from the serial port until the current message from
the device is complete. The theory is stopping mid-message
might leave the device in a confused state where it
keeps putting the same message on the wire waiting for
the other side to receive it correctly. This should
not impose any additional delay since an active stream
means the loop isn't waiting for an IO timeout