Experimental caching idea
To try this out, you need to have corepos/class-cache 1.1+
installed.

The basic idea here is to take classes that POS uses frequently
and combine them all into one big file. The potential savings
here is that instead of hitting the hard drive once for each
class there's one hard drive operation that brings in a whole
raft of classes.

The current setup will immediately caches all the Parsers.
Time from entering a command to seeing a response is a big part
of perceived performance. The system then tracks how frequently
the autoloader has to fetch other classes. If a class is requested
five times in less than two seconds it gets added to the cache.

One downside here is code changes won't always display instant
results. The cache is only refreshed on sign out + in (or otherwise
going to the root index.php script). Another downside is that
error messages will refer to line numbers in the cache file
rather than the original source. Because the cache file changes
in response to usage the logged line number isn't guaranteed
to correspond to the current cache file. Logging with a stack
trace would help some but that's not available on fatal errors.