JS => pos.exe error handling
When the browser uses an AJAX call to communicate with
pos.exe and the connection fails, the javascript error handler
still assembles and submits a response (xml-resp). However, if
no text can be extracted from the XMLHttpRequest the value that's
submitted is just the empty string. On the PHP side the empty
string was not properly distinguished from no response at all.

The primary fix here is comparing the submitted value against
false rather than the empty string. This way the transaction
is always finalized which tidies up PaycardTransactions an takes
the user to an error screen.

The LaneLoggger bit is there to hopefully track down more info
about what may have gone wrong. A localhost HTTP connection really
shouldn't just drop at random. Logging more real-world examples
might make it possible to provide a better, less generic user
facing error in this case.

The emv.js changes to always submit a non-empty value are just
to prevent a PHP-side warning that occurs when trying to parse an
empty string as XML.