Add login to speed profiling
This has to pull a time slightly early because kick()
ends the session in case anything goes wrong with the
printer hardware. This shouldn't matter a ton though.
Time it takes to finish the script should be things out
of CORE's control - how long it takes to write to the
printer, how fast PHP/webserver are running, etc.