Support for P2PE w/ gift cards
The service actually declines the encrypted gift card transactions
but it returns the full card number. Simply re-submitting the card
number returned by the service results in an approval. This adds
minor overhead but is presumably going to be fixed at some point.

XmlData is also replaced with BetterXmlData for more reliable
parsing. get() vs get_first() with the former was always a little
wonky.