Revive old TCP print idea
Definining a printer port setting that starts with tcp://
(i.e., tcp://localhost:9100) will open a TCP connection
to the host+port, send the raw receipt data, and close the
port. I have no idea which printers this will and will not
work for but is simple enough to try.