Cashback handling tweaks
paycard_amount needs to be the total amount to charge.
If cashback changes, reset to amtdue + cashback. This means
if someone wants to pay a partial total AND get cashback the
cashback needs to be entered first, but if the due amount
doesn't reset then entering consecutive cashback amounts
will cause more problems as the total keeps growing