Offer HTTPS on login screen if configured
I added a config setting for an HTTPS domain. For HTTPS to
work *correctly* the server needs a domain name, at least on
the local network, instead of just an IP. The login page is
where someone is most likely to encounter a user-facing message
from the browser about security so that's where I added the
switch link and the config setting.