Intercept end-of-transaction idea
This uses a PreParser so that inputs that might end the transaction
can be identified *before* any Parser that would consume the input
runs. The plugin's PreParser simply appends a prefix that the
Parser strips back off. Control then flips to a plugin page for
input before returning to the main page and repeating the originally
intercepted input.