Overhaul of weighted margin in trans_archive data
The way the percentages of sales were calculated for this
just weren't right. There were some basic flaws in a multistore
setup that didn't keep their sales properly separated, but
the weight averaging just didn't work.

The revised setup walks through the weekly data, one item at a time,
to make a blended average that counts 100% of the first week,
95% of the second week, 90% of the third week, etc. This is slower
and may add a few minutes of run time, but gives more sensible
results. The $totalCache stuff is to cut down on redundant queries
as many of the percentages wind up needing the same denominators.