Experiment w/ job queues and XLS batch
Push some extra changes in the uploaded file into the
job queue and let the those updates happen separately
in the background.

There's not necessarily a practical reason to do this.
It's just an opportunity to test the queue idea more
and see what happens to stability and resource use
when pushing more jobs into it.