New default prodFlags out-of-stock, discontinued
These are supplemental to inUse as an active/inactive toggle

An out-of-stock item might still be actively inUse but not
currently present. Since automation here revolves around
purchase orders an item experiencing vendor out-of-stocks might
get flagged as out-of-stock in POS before the on-hand quantity
gets down to zero.

A discontinued item is sort of the inverse. It might be
still physically present but is in the process of going away
and becoming inactive.