Driver port re-negotiation
Occasionally the pos.exe driver crashes or exits in
such a way that the TCP and UDP ports it uses to
communicate remain bound to a non-functional, unkillable
zombie process. Trying to start pos.exe again at this
point crashes immediately because the ports its trying
to bind to aren't available.

The idea here is to catch the exceptions that happen
when a port isn't in use and keep trying higher numbers
until an available port is located. If it uses a non-standard
port it also sends a file-based message to POS telling it
which port(s) to use for communication.

This could get out of sync if pos.exe crashes & reloads on
non-standard ports than the user closes & re-opens the browser.
That's probably fine. Closing & restarting both the browser
and visible POS driver window should bring things back into
sync. Most of the time this crash behavior causes users to
reboot. If the initial crash is handled smoothly but later
confusion leads to a reboot nothing's really been lost.