ScanBox support
This is a Socket Mobile browser app. It's not super
capable but can insert its scans to a specific element
by id.

This adds a hidden input with id "socketm" and uses
an object setter to intercept the scan data from the
socket mobile and insert it into the normal handler
functions. I doubt there's complete parity with the
Linea support but this should handle a lot of cases
with tweaking individual pages.