Legacy $new_item out, more expansive comment
`$new_item` in `write_items_to_scales()` was deadwood, never used.

The original comment in `getItemLine()`, left in place except for a typo fix, I don't really understand, especially "'updating' an item that doesn't actually exist on the other side of processing".  The new comment is related but may not replace it.