Log survey issuance
The sequence of events means when a survey is issued the
log record will wind up in the *next* transaction since
data has already been rotated. In practice I don't think
this matters. The most common usage will just be querying
upc='CXSURVEY' to get the total number issued. If we need
to map the log record back to the correct transaction it
does contain trans_num in the description