Cron task for re-writing repack quantites
When POS interprets by-weight scale labels it only substitues
a "real" quantity if dividing the label price by the per weight
price, rounding, and re-multiplying matches the label price. If
there's a discrepancy it gives up and uses quantity 1 to avoid
conflicting with what the label might say.

On the backend having a quantity that somewhat close is better
than just a default 1 so this task adjusts quantity and pricing
columns for the previous day. The "total" value will not change.