BUGFIX: Databar coupon extra rules field
When present the 2nd required item fieldset has a value
indicating how to combine the 2 or 3 required items.
Implementation wasn't reading this value so everything that
follows is misaligned when a 2nd requirement is present.

I also changed the interpretation for the rule value 3.
It currently allows the coupon to be used if *any* of the
required items is present. I think the purpose of this is
to allow for situations where eligible items could have
different prefixes.