Avoid "orphaned" SKU when saving product
assuming the following:

- UPC 07430500132 exists in products but has no vendorItems record(s)
- UPC 07430500116 does not exist in products, but
- there is a vendorItems record with SKU 12345 and UPC 07430500116

with this commit, user is allowed to edit 1st item and assign it SKU
12345, even though that is associated with a different UPC.  since
there is not a corresponding products record, the vendorItems record
is considered "orphaned"

without this commit, user is not allowed to assign SKU to 1st item,
b/c of the orphaned vendorItems record.

the code change basically ignores orphaned records, so they do not
prevent user from taking desired action in this particular case.