SNAP + Tax on House Coupons
This is a narrow solution where discount type P+ checks
the discount items to find the minimum applicable tax
rate on foodstamp-eligible items. If found, that tax rate
ID is used in the coupon's transaction record, and the
foodstamp flag is also set for the coupon record.

This is only addressing a specific tax forgiveness issue.
Doubtless other questions about how coupons interact with
tax still exist