Allow for token based authentication
User can generate one or more tokens for use as an
alternative logging in. To use this functionality,
put the token in an Authorization Bearer HTTP
header.

For simplicity's sake, there are no permissions
tied to any given token. Each token has the same
permissions as the user it belongs to.