Backend lane detection
See also: 4721f62816eed67d243198600fa3fbea45a5568d

The basic idea here is to leverage the fact that the
lane's hardware driver already includes a mini UDP
messaging service. The backend broadcasts a message
and the lanes respond with their IPs. If this proves
to be lightweight enough, it could be used for
real-time checking before an operation that needs
to communicate to the lanes instead of relying on
periodic polling and tracking online/offline status.

In theory this could also be used to eliminate the need
to configure a set of lanes, although I think it'd require
much more security in the messaging to be a good idea.
Right now it'd be very easy for something else to answer
the broadcast and pretend to be a lane.