Versioning for FanniePlugin
FanniePlugin now includes a $version property that controls
where the plugin's settings are stored.

The default version is 1, and settings are stored in config.php
under $FANNIE_PLUGIN_SETTINGS as is currently typical.

Version 2 plugin settings are stored in the database table
PluginSettings instead.

The plugin configuration page supports the different versions
and will migrate settings to the database automatically if
a plugin's version has increased. I think it'll probably do the
reverse, too, although I haven't tested that.

This commit also add another FanniePlugin property $settingsNamespace.
While this does sort of work and settings get stored in the database
with the namespace attached as a prefix, the way it's currently
implemented on the configuration page **does not** prevent naming
collisions. A namespace might be somewhat useful when querying
PluginSettings, but how they're handled on the configuration
page will need more work before namespaces are a proper feature