Big cleanup of deprecated/removed stuff in bundled adodb
I deleted a fair amount of stuff that wasn't relevant
or ever used by CORE anyway.

I left mysql and mssql in place even though these were removed
in PHP 7.0. CORE has historically supported them, and
the official system requirements still list the minimum
PHP version as 5.5. In the case of mysql there might
genuinely be an install somewhere relying on this still.

I also left the ibase stuff since it's not in PHP proper
anymore but appear to be available via PECL