Fix for #1088
Namespacing of settings should only take place if
the setting has a current value. Settings for disabled
plugins will not have values and should be ignored.

In the course of verifying this eliminates the warnings,
I found that calling settingChange() on enabled plugins
needs to be postponed and called *after* the new
settings have been saved to the database