Add doctrine/dbal as substitute for ADOdb
This alters to SQLManager to
1. Create connections using doctrine/dbal if present
2. Use those connections for all functions

It can be disabled by configuration. Those checks aren't
currently applied correctly so that CI tests can run,
but if this gets merged those will be changed so it
defaults to using ADOdb as normal.

This could someday solve the problem of bundling in
a pretty old copy of ADOdb.

I did remove a couple functions, affectedRows and dataSeek,
each of which had literally one use. Altering those spots
seemed easier than finding a way to make doctrine
mimic those functions.