Mixed price BOGO handling
When two items in a BOGO set have differing prices, make
the cheaper one free. This does NOT rexamine the entire
transaction. It only compares the current item to the item
it's being paired with. So the ring order is still relevant
when dealing with multiple pairs, but overall it should be
a fair bit more consistent.