Add initial support for "Poser" concept (#1074)
* Add initial support for "Poser" concept

so far this only lets Fannie plugins exist in Poser

* Prefer plugin search directories from Poser over CORE

also use a better naming convention within Poser

* Expose composer config value in UI

I'm not sure if "themes" is the ideal place but it is
customization related. Warns if the provided path doesn't
exist or isn't absolute. Could probably later do more
validation that the directory is in fact recognizably
a Poser instance.

* Allow direct loading of Poser classes w/ namespaces

* Lane-side basic Poser implementation

* Fix syntax bug

* Only attempt autoload of lane plugins from poser, if dir exists

* Rename folder to `office_plugins`, for consistency

Co-authored-by: Andy Theuninck <andy@gohanman.com>